+++ PayPal/src/main/java/com/braintreepayments/api/PayPalLineItem.java
@@ -1,5 +1,8 @@
 package com.braintreepayments.api;
 
+import android.os.Parcel;
+import android.os.Parcelable;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.StringDef;
 
@@ -9,7 +12,7 @@ import org.json.JSONObject;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
-public class PayPalLineItem {
+public class PayPalLineItem implements Parcelable {
 
     /**
      * The type of PayPal line item.
@@ -182,4 +185,45 @@ public class PayPalLineItem {
 
         return new JSONObject();
     }
+
+    PayPalLineItem(Parcel in) {
+        description = in.readString();
+        kind = in.readString();
+        name = in.readString();
+        productCode = in.readString();
+        quantity = in.readString();
+        unitAmount = in.readString();
+        unitTaxAmount = in.readString();
+        url = in.readString();
+    }
+
+    public static final Creator<PayPalLineItem> CREATOR = new Creator<PayPalLineItem>() {
+        @Override
+        public PayPalLineItem createFromParcel(Parcel in) {
+            return new PayPalLineItem(in);
+        }
+
+        @Override
+        public PayPalLineItem[] newArray(int size) {
+            return new PayPalLineItem[size];
+        }
+    };
+
+    @Override
+    public int describeContents() {
+        return 0;
+    }
+
+    @Override
+    public void writeToParcel(Parcel parcel, int i) {
+        parcel.writeString(description);
+        parcel.writeString(kind);
+        parcel.writeString(name);
+        parcel.writeString(productCode);
+        parcel.writeString(quantity);
+        parcel.writeString(unitAmount);
+        parcel.writeString(unitTaxAmount);
+        parcel.writeString(url);
+    }
+
 }